requirements.txt
flask
minimongo
feedparser
dateutils
nose
ipdbplugin
py-bcrypt
Hamlish-Jinja
flask-login
flask-fillin
